% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{bq_connect}
\alias{bq_connect}
\title{Connect to BigQuery using the active (or specified) module/env}
\usage{
bq_connect(module = NULL, env = NULL)
}
\arguments{
\item{module}{Optional module override.}

\item{env}{Optional env override.}
}
\value{
A lazy \code{dplyr} tbl.
}
\description{
Calls \code{bigrquery::bq_auth()} if not already authenticated.
Returns a pre-joined lazy \code{dplyr} tbl.
}
