% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{load_rules}
\alias{load_rules}
\title{Load rules from the active environment file}
\usage{
load_rules()
}
\value{
a tibble of rules
}
\description{
A function that creates a tibble containing all the individual
rules.  These should be in an Excel file
}
