% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_reader.R
\name{reload_config}
\alias{reload_config}
\title{Reload and revalidate the config from the stored YAML path}
\usage{
reload_config()
}
\description{
Reload and revalidate the config from the stored YAML path
}
