% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{rules}
\alias{rules}
\title{Get the currently loaded rules}
\usage{
rules()
}
\value{
The currently loaded rules, or \code{NULL} if no rules have been stored.
}
\description{
Retrieve the rules stored in the active package state.
}
\seealso{
\code{\link[=load_rules]{load_rules()}}, \code{\link[=reload_rules]{reload_rules()}}
}
