% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Qctype_mapping}
\alias{Qctype_mapping}
\title{QC Type Mapping Configuration}
\format{
An object of class \code{list} of length 20.
}
\usage{
Qctype_mapping
}
\description{
A nested list mapping QC types to their expected R classes and NA requirements.
}
\details{
A named list with X elements
\itemize{
\item check_type
\item is_na_ok
}
}
\keyword{datasets}
